feat: expose context brief from plugin hooks
Share the OMH context brief builder across CLI, plugin tools, and pre-LLM hooks so Hermes hosts can receive a structured omh_context_brief/v1 alongside the bounded text context.

Constraint: Keep Hermes-facing context metadata-only and advisory without echoing raw prompts.

Rejected: Leave omh_context as a tool-only surface | wrappers need hook-time context before generic image, file, search, or coding tools run.

Confidence: high

Scope-risk: moderate

Directive: Treat hook context briefs as routing/status guidance only, not workflow execution, generation, dispatch, verification, review, CI, merge, or delivery evidence.

Tested: PYTHONPATH=tests uv run python -m unittest discover -s tests -v; uv run python -m src.cli docs workflows --check; uv run python -m src.cli release product-readiness --json; uv build; git diff --check

Not-tested: Live external Hermes host rendering of the structured hook payload.
